[7185] Amended withdrawal attributes to be versioned
